+++ lib/models/TransactionError.dart
@@ -5,4 +5,19 @@ class TransactionError implements Exception {
 
   @override
   String toString() => this.message;
+
+  static String formatError(dynamic error) {
+    try {
+      if(error["errors"] != null) {
+        final errorList = error["errors"] as List<dynamic>;
+        return "${errorList.first["message"]}";
+      }
+      if(error["message"] != null) {
+        return error["message"];
+      }
+      return "An unexpected error occurred. Please try again.";
+    } catch (error) {
+      return "An unexpected error occurred. Please try again.";
+    }
+  }
 }