+++ lib/models/requests/customizations.dart
@@ -1,3 +1,5 @@
+import 'package:flutterwave_standard/utils.dart';
+
 class Customization {
   String? title;
   String? description;
@@ -7,10 +9,11 @@ class Customization {
 
   /// Converts instance of Customization to json
   Map<String, dynamic> toJson() {
-    return {
+    final customization = {
       "title": this.title != null ? this.title : "",
       "description": this.description != null ? this.description : "",
       "logo": this.logo != null ? this.logo : ""
     };
+    return Utils.removeKeysWithEmptyValues(customization);
   }
 }